Refactor bigEndianEncode and getBloomFilterBit
