+++ server/differentialPrivacy/dpAnalysisController.js
@@ -3,6 +3,8 @@ const http = require('http');
 const fs = require('fs');
 const md5 = require('md5');
 
+const { bigEndianEncode, getBloomFilterBit } = require('./dpUtility');
+
 const {
   BLOOM_FILTER_SIZE,
   NUM_HASH_FUNCTIONS,
@@ -23,28 +25,6 @@ const sampleCandidateStrings = ['v1', 'v2', 'v3', 'v4', 'v5', 'v6', 'v7', 'v8', 
   'v151', 'v152', 'v153', 'v154', 'v155', 'v156', 'v157', 'v158', 'v159', 'v160', 'v161', 'v162', 'v163', 'v164', 'v165', 'v166', 'v167', 'v168', 'v169', 'v170',
   'v171', 'v172', 'v173', 'v174', 'v175', 'v176', 'v177', 'v178', 'v179', 'v180', 'v181', 'v182', 'v183', 'v184', 'v185', 'v186', 'v187', 'v188', 'v189', 'v190',
   'v191', 'v192', 'v193', 'v194', 'v195', 'v196', 'v197', 'v198', 'v199', 'v200'];
-
-// Encodes an input integer to a 4-byte big-endian string.
-function bigEndianEncode(value) {
-  let result = '';
-
-  // Value must be an integer (cohort number)
-  value = parseInt(value);
-
-  for (let i = 24; i >= 0; i -= 8) {
-    const byte = (value & (0xFF << i)) >> i;
-    result = result.concat(String.fromCharCode(byte));
-  }
-
-  return result;
-}
-
-function getBloomFilterBit(inputString, cohortNum, hashNum) {
-  const toEncode = bigEndianEncode(cohortNum) + inputString;
-  const hash = md5(toEncode);
-
-  return parseInt(hash.slice(hashNum * 2, hashNum * 2 + 2), 16) % BLOOM_FILTER_SIZE;
-}
 
 // Generate the counts file containing the total number of reports and bit sums
 // for each cohort.
@@ -113,7 +93,7 @@ function generateParamsFile() {
 // Given an Array of candidate strings, using aggregated data in Redis,
 // returns a Promise that is resolved with an Array of detected strings.
 function performDPAnalysis(candidateStrings) {
-  
+
 }
 
 // Exports
